Switched to yarn as npm is broken
Ref:
https://github.com/npm/npm/issues/19019
https://github.com/nodejs/node/issues/16649

npm team member said "The “real” fix would be waiting for the next npm
release." Yet I don't want to wait at all!

Seems some of npm's dependencies are not compatible with node 9
https://github.com/npm/npm/commit/c851bb503a756b7cd48d12ef0e12f39e6f30c577 (Bumping node-tar)
https://github.com/npm/node-tar/commit/ea90c6cec64044a45448b90bd363c4bf603aff9f (Bumping minipass and minizlib)
https://github.com/isaacs/minizlib/commit/a40d6ce1c6465e7488e522a5f8e5373b27d45152 (Update zlib API usage for node 9)

And it's rather stranger for a scripting engine that crashes for a known
misuse. A JavaScript exception should be thrown instead.