Include generated version in VCS
The VERSION text from the PKGBUILD template is just a placeholder.

The docs state:

> To use pkgver(), you still need to declare the pkgver variable with
the most recent value.

The value is generated there by makepkg, it just needs to be included in
the PKGBUILD and committed to source control.

To apply this in AUR the pkgrel also needs to be bumped, because there's
no upstream change.