Initial commit
This is the unified "xmr-stak". The PKGBUILD, as configured, builds only
the CPU miner for monero without SSL and the HTTP server and disables
the default donation. If you want more features - just flip the
corresponding cmake option from OFF to ON and add the necessary package
as a dependency:

https://github.com/fireice-uk/xmr-stak/blob/master/doc/compile.md

If you want to donate to the developers - comment-out the prepare()
step, or you can directly send donations as it is described here:

https://github.com/fireice-uk/xmr-stak#default-developer-donation

The necessary configuration will be created on first run.

You can raise the number of huge memory pages and the limit on the
amount of memory the user can pin like this:

sudo echo 'DefaultLimitMEMLOCK=2G:2G' | sudo tee -a /etc/systemd/system.conf
sudo echo 'vm.nr_hugepages = 128' | sudo tee /etc/sysctl.d/xmr.conf
sudo systemctl restart systemd-sysctl.service