Update to v2.0
This is a big update. It changes the implementation from a Python script
to a compiled binary written in Go.

As a result, there are lots of changes to the PKGBUILD. I am not well
versed in working with Go and have made lots of hacks in writing the
PKGBUILD just to get it to work.