Upgraded to upstream version 2.10
This is a major change. We are now building the Arch package from
Linux-specific binaries hosted on github instead of building from
source. This gets around some recent breakage with GHC and Arch. It also
means this package creation process is faster and has a much smaller
drive space footprint.