Disable XZ compression
Because this package very big, it takes a very long time
to compress with XZ compression. By disabling it, and
using only pkg.tar file format, the size of the tarball
increases in 200MB (from 400MB to 600MB) but the
compression time decreses drastically.
It seems to be a good trade-off.