fix errors during build
- I decided to turn off binary compilation of this package, because some
of the source file should not be compiled at all (boilerplate
directory).
- I deleted some optional dependencies since there is no reason for them
to be installed by a regular user (they are used mainly for purpose of
documentation building). Developer should clone Abjad from github or
install it via pip with all the optional dependencies instead of using
this aur package.