Update optdepends and refresh pkgver
Mostly follows v4 of the mkosi package. Reinstates edk2-ovmf, because it
seems to be available (again?) from AUR, and as far as I can tell it
contains the necessary files while ovmf doesn’t, but I’m not sure about
this at all anymore. I hope this will be cleared up eventually :/