depends: switch libva-git to libva
The intel-media-driver is based on a specific libva commit that
comes between the 2.0.0 and 2.1.0 releses. Now that libva 2.1.0
was released, we can use the regular libva package from the
official repositories.

Currently it seems to work fine with libva 2.1.0. Since this is
the development (-git) package, maybe it will be needed to switch
back to libva-git in the future if upstream requires libva
development code in future changes. Let's follow on.