lxd - update file/process limits
The lxd package currently uses, the default file descriptor limit
(1024). This reduce the usefullness of this package if users want
to run larger containers. The Arch Linux wiki suggests to remove
this limits. The ubuntu package also ships with much higher limits,
than the defaults. I updated this package to match the Ubuntu
defaults.