Change the version of urwidtrees in setup.py to >=1.0.2
The urwidtrees upstream official current release is 1.0.2, the one in
pypi is an untagged release.

For stig the only reason to use the 1.0.3dev0 is because of a packaging
bug fixed in the git master of urwidtrees and pip, this don't affect us.

This allow the use of the community/python-urwidtrees or the versions in
aur.

References:

https://github.com/pazz/urwidtrees/issues/37
https://github.com/rndusr/stig/pull/7#issuecomment-275389365