Revert "Fixed PKGBUILD"
This reverts commit 04f631afce07aeb73041c87b524a393ec82cd6b8.

In no sane world does this fix anything at all, instead a number of
things have regressed. Specifically, the badly designed use of bash in
the systemd service, the fact that an architecture-dependent package
was casually switched to lie and claim it is an "any" package, moving
away from the upstream naming for the service, and not cleaning up after
being stopped.