protonmail-bridge - add gnome-keyring dependency
Apparantly the ProtonMail bridge also has a dependency on gnome-keyring.
If gnome-keyring is not installed, it would not be possible to add an
account. It may be possible that other services inplementing the
org.freedesktop.secrets specification/gnome-keyring apis would also
work, but I couldn't find another package inplementing this
specification in the official repositories Arch Linux, and
gnome-keyring does not provide a virtual package to depend on. Because
gnome-keyring has (implicit) dependencies on gst-plugins-base-libs, gtk3
and libpulse, we also remove those dependencies.