snapd - add .tar.gz extension to source tarball
Previously the source tarball was downloaded to snapd-$pkgver. This is
not a really big problem, since tar is still able to extract the file.
However it might create some unnecessary confusion about the file type,
and there is no reason not to add it. Therefore, since this commit
makepkg downloads the source code to snapd-$pkgver.tar.gz.