Add version to downloaded files
There was a problem that the package was not updated if the downloaded
files were cached (because the file exists and pkgver() reads the
version from the old file).

Now the name of the downloaded file will include the version. Checksums
are still skipped, so the package will always build, but the version in
the filename might be wrong.