Add "java-openjfx" package dependency
This ensures that the required JavaFX runtime which is required to run
the application. Also as of Java 11 the official Oracle JDK won't
include JavaFX anymore.
With the Java Platform Module System in place since Java 9, Oracle
decided that it now makes sense to decouple JavaFX from the JDK, in
order to make it available as a separate download.