Replace cp by install
and thereby remove the need for explicitly creating the directory to
copy the docs in, which might have been pointless in the case of no doc
existing.

Also, replace the short-circuiting using `&&' by an if-statement in
order to prevent package() from failing.

The changes do not affect the resulting package.