Use sources instead of binaries
Binary packages are now to be found as dict-freedict-*-bin.

Also, update to version 0.3.7.

Furthermore, replace `cp' by `install' and thereby remove the need for
explicitly creating the directory to copy the docs in, which might have
been pointless in the case of no such file to be copied existing.