Replace `cp' by `install' and specify make target
Replacing `cp' by `install' removes the need for explicitly creating the
directory to copy the docs in, which might have been pointless in the
case of no such file to be copied existing.

Explicitly specifying the make target appears to be required now; also
the location of the built dictionary files has changed.

Finally, replace the short-circuiting using `&&' by an if-statement in
order to prevent package() from failing.