Update to latest upstream version 1.7.0
The .deb source file is packaged using another packager tool instead of
install4j. It comes with a bundled JRE Java runtime so the dependency to
the metagroup "java-runtime" is currently not necessary anymore. The
bundled JRE should be used until the compability with the latest JRE
9/10 version has been approved.

Another change is that the application icon is not bundled anymore so it
has been added to the package itself.

The also .deb file uses another format
resulting in a "data.tar.xz" archive instead "data.tar.gz" so the "z"
flag of "tar" has been changed to "J" to use "ar x <NAME>.deb" and
"unxz" to uncompress the file and data archive.