Fix PID file location on upgrade
post_upgrade() is usually run under sudo, so $XDG_RUNTIME_DIR is not set
and $(id -u) is pretty much useless.  So if neither gives us a usable
path, try /var/run/user/$SUDO_UID as a last resort.

Note that in addition, the /var/run path was wrong.  It is
"/var/run/user", not "/var/run/users".

Good thing this happened on the first version so probably nobody has run
post_upgrade() so far.