Do not force the use of 'iHD' libva driver by default
This will let the user choose which libva driver to use. No other
libva driver package on the official repositories force the setting
of the installed driver by default. This makes sense because the
user can install this additional iHD driver without modifying any
settings that he might have previously configured.

Configuration is done by setting the 'LIBVA_DRIVER_NAME'
environment variable. In the case of this package, it
must be setted to 'iHD'. See the Wiki page on Hardware
video acceleration (VA-API sections) for details. An
informative message is shown during install and upgrade.