makedepends: switch libmfx to intel-media-sdk
intel-media-sdk, the open source version of
intel-media-server-studio, has now reached an acceptable stable
release that can be used here to provide libmfx.

This brings 2 important results:

1) The experimental state can be removed from the Intel QSV
   support, since intel-media-sdk also provides libmfxhw64
   and plugins for hardware accelerated decode and encode.

2) Finally we can get rid of the unsecure rpath in ffmpeg
   binaries!

intel-media-sdk is now a makedepend, instead of an optdepend
as previously. If fact, I should have switched it to
intel-media-server-studio in optdepends some time ago when
it has changed to the open source version, but I letted this
way while waiting for an acceptable stable release of, which
taked a while to happen.

Users should remove the libmfx{-git} package and stay only with
intel-media-sdk{-git} package, since the former is the complete
solution with the dispatcher library (libmfx.a), libmfxhw64 and
plugins.

Note1:
------
For the proper use of Intel QSV in ffmpeg the user must set the
vaapi/libva driver to 'iHD' through the 'LIBVA_DRIVER_NAME'
environment variable. The 'iHD' vaapi/libva driver is provided by
intel-media-driver{-git} package, which at the current time is
already a dependency of intel-media-sdk{-git} package. Setting
the vaapi/libva driver to 'iHD' can be done, for example, by
uncommenting the proper line in '/etc/profile.d/intel-media.sh'.

Note2:
------
intel-media-sdk is currently at a pre-release state. This can
lead to bugs when using Intel QSV.

Note3:
------
Not all Intel platforms are supported by intel-media-sdk. At the
current time, it claims to support only 5th and 6th Generation
Intel® Core™ processors (Broadwell and Skylake). For details,
please see the MediaSDK and media-driver documentation:
- https://github.com/Intel-Media-SDK/MediaSDK/
- https://github.com/intel/media-driver/

Note4:
------
Intel QSV is available in x86_64 systems only.

Also in this commit:
--------------------
- depends: moved 'libbs2b' and 'sndio' from the AUR section to
  the official repositories section, since they are now at the
  [community] official repository.