Move intel-media-sdk from makedepends_x86_64 to depends_x86_64
Since intel-media-sdk is needed not only for building with
Intel QSV support but also for Intel QSV to work during runtime, it
makes sense to stay in depends as a needed component. Forgot to do
it in previous commit. Another option instead of placing it only in
depends would be to place it in makedepends + optdepends because
ffmpeg runs without intel-media-sdk installed (libmfx.a is a static
dispatcher library that calls other sdk components at runtime).
Currently I'm sticking with the depends-only solution.