Fix build: add fixes for cuda 9.1, needs gcc5
Unfortunatelly, the current available stable version of caffe2
does not build with the current cuda 9.1 and gcc6 from the
official repositories. This prevented the package to be built
for a long time.

Details here:
https://github.com/caffe2/caffe2/issues/1459
https://github.com/caffe2/caffe2/issues/1636

But it is possible to build with gcc5 from the AUR.

This commit adds build fixes for cuda 9.x, including the use
of gcc5. Be warned that that gcc5 from the AUR takes a lot of
time to compile.

Important notes
---------------
1) The Caffe2 source code moved to PyTorch repository, but
   currently there is no stable version containing Caffe2 there.
   It is yet to be released.
2) The current available stable version of Caffe2 does not have
   python3 support. Python3 support is available on the git
   master branch. If you want python3 support right now, please
   use caffe2-git or caffe2-cpu-git package.