propagate dependency version change
The version actually was restricted to >=0.16.0 in configure.ac before
without this being noted anywhere.

In contrast to other variables like depends, which can be set in a
package function, or pkgver, for which a dedicated function exists, an
automatic detection of this restriction is not possible, as it has to be
known before preparation even starts.