Fix build: use gcc5. Switch to python3. Change source repo.
Fix build: use gcc5
-------------------
Unfortunatelly, the current git master branch of caffe2
does not build with the current cuda 9.1 and gcc6 from the
official repositories. This prevented the package to be built
for a long time.

Details here:
https://github.com/caffe2/caffe2/issues/1459
https://github.com/caffe2/caffe2/issues/1636

But it is possible to build with gcc5 from the AUR. Be warned
that gcc5 from the AUR takes a lot of time to compile.

Switch from python2 to python3
------------------------------
Upstream now provides python3 support at git master branch.

Change source to pytorch repository to match upstream changes
-------------------------------------------------------------
Caffe2 source code moved to PyTorch repository.

References
----------
https://github.com/caffe2/caffe2/issues/315#issuecomment-350974915
https://github.com/caffe2/caffe2/commit/a062b05d70c75a6005c6e76109a1dde482729a53#diff-04c6e90faac2675aa89e2176d2eec7d8