Initial commit
Extract the jar file under Graal’s JAVA_HOME, and then adjust
permissions and symlinks according to the respective META-INF files
(since the zip format is not guaranteed to support them).

This is enough to make R work (assuming /usr/lib/jvm/default/bin/ is in
your PATH), including polyglot use (R --polyglot, eval.polyglot()), but
apparently not to enable other Graal languages to use R. I’m not sure
why that is, or if that’s even a fault of this package (I seem to get
the same result when installing the package with `gu` directly).