Initial commit
PKGBUILD copied from fastr and then slightly adjusted.

This seems to be sufficient to make Ruby work on its own and as a
polyglot language for R. However, attempting to launch Ruby with the
polyglot option results in an error:

    $ /usr/lib/jvm/java-8-graal/bin/irb --polyglot
    ERROR: Unrecognized argument: --ruby.launcher=/usr/lib/jvm/java-8-graal/jre/languages/ruby/bin/ruby. Use --help for usage instructions.

Again, I get the same error when installing via `gu`, so I assume it’s
not the PKGBUILD’s fault.