Update git submodules. Change pkgver method.
Update git submodules to match latest upstream changes.

Note that 'third_party/nccl' is not a real git submodule
and thus it does not need to be cloned by makepkg (nccl
source code currently comes bundled in 'third_party/nccl'
when cloning the pytorch git repository).

$pkgver now uses the caffe2 version from the file
'caffe2/VERSION_NUMBER'. Since this version does
not coincide with the pytorch version mumber, the
revision number is now the total number of revisions
in the pytorch git repository.