Initial commit
This package aims to provide a way to install jpf
from the binary snapshots hosted on the project's website.
Due to a few deprecations in recent versions of Java 8,
building the sources have become impossible on some Linux setups,
and this package has been created to get around this issue.