Disable rocksdb and aten
AUR package rocksdb is currently broken. It does not build. It
also suffers from poor package designing, as it produces different
builds depending which packages the user have installed (for
example, intel-tbb and jemalloc - it should explicitly enable them
at depens and build(), or explicitly disable them at build(), but
the author prefers to let it ambiguous, and this causes problems).