Change destination dir of upstream source to fulfill makepkg needs
makepkg does not allow the destination directory of upstream source
to have the same name of the package anymore. This was previously
allowed, but now it will produce an error saying that the directory
is not a clone of the upstream source. Existing directories on the
old allowed schema will still work, but clones of upstream source
will now produce this error.