Update ChangeLog for 4.015
The Fontlog.txt on GitHub [1] has some extra intermediate versions that
were never published on vollkorn-typeface.com as far as I’m aware (nor,
consequently, as versions of this package). I guess the best way to make
sense of them is to squash their changes into the next released version.

[1]: https://github.com/FAlthausen/Vollkorn-Typeface/blob/master/Fontlog.txt