Add woeusb-git to conflicts array
This package obviously conflicts with woeusb-git, so I added it to the
provides and conflicts arrays. Technically, woeusb-git is suppossed to
conflict with woeusb. [1] That is why the conflicts/provides arrays were
empty. On the other hand, woeusb-git is far more popular at the moment,
so I updated these arrays anyway.

[1] https://wiki.archlinux.org/index.php/PKGBUILD#conflicts