Remove optdep for libtxc_dxtn; modernize PKGBUILD
libtxc_dxtn is no longer necessary, and is no longer available on the
AUR, so remove the reference to reduce confusion.

Remove the .install file because gtk2 now seems to use fallbacks if the
engine is not available. Use bsdtar instead of unzip to avoid the
makedep. Use exec in .sh since the bash process is useless while
playing.