Update to 0.9.1(-emotefix)
A number of things have changed in upstream since 0.8.7, notably:
* The license for the program as a whole has changed from MIT to GPL3
* It's now using gradle instead of ant for building

For this and other reasons, a lot of the patches that were previously
included is being dropped here on out, unless it turns out that they're
desperately needed. The only thing being patched now is a variable that
determines whether to do version checks (which is being disabled, as per
upstream's recommendation), all other code is the same as upstream's.

One thing this means is that the Chatty configuration will now live in
`~/.chatty` instead of `~/.config/chatty` which may "break" some users'
experience. I will post a note on the AUR page about this.