Download from nodejs registry
Previously this package downloaded from the GitHub release. This meant
we had to use a quite buggy build system. If we download from the nodejs
registry, we can run 'npm install -g' and let npm do the packaging for us.
Since this tool will be available from npm, this should keep working for
the forseeable future. Even if the packaging structure in GitHub changes.