Fix location of LICENSE file
With -t, the target is always treated as a directory, including the last
component. We could fix that for this package by not specifying the last
component of the target (remove the LICENSE), but for consistency with
other Graal packages, were we sometimes have to specify the last
component because the source file is not named LICENSE but something
like LICENSE_GRAALPYTHON, I prefer swapping the arguments back into
SOURCE DEST order, and employing the -T option just in case.