Added device description files.
I have decided that it is against the interest of the community as a whole to
not include the device description files in this package and to have two
separate packages potentially causing any confusion. As such, I have included
the device description files in this PKGBUILD to be installed. I apologize for
the incident.