Download file from mirror
It seems that opendesktop.org changed something on their platform that currently does not allow to download files from a script.

The download link in their platform dynamically generates a download hash, so it is not possible to use the previous links.

Until I manage to get this right, a mirror URL will be used.