download waf as a makepkg source
This allows building without network, caching the waf download even when
doing a clean build, and is cleaner semantically in general.

We still run mpv's bootstrap script, but we now tell it not to attempt
the download. It will simply check that the waf version it expects is
present. If that check fails, it will bail and the PKGBUILD will have to
be updated.