added streamlink optdep and fixed versioning
for now the version has been reverted to the old way of just taking the
commit hash, if later a stable tag format is established or a better way
of incorporating tags into version is found it'll be switched to that.
but this will at least stop it from incorrectly assuming version on aur is
always newer than local.