Move to /opt and included license
In order to adopt a more KISS approach, all the code
moving the files from /opt to /usr directories were
remove.  As a consequence the PKGBUILD is much more
simpler and readable and there is not pratical
difference.

Also LICENSE file was obtained from documentation and
installed into the package, which is a requeriment
for software with custom licenses in Arch.