sparforte: use SOURCE_DATE_EPOCH and other fixes
This disables mysql support for APQ since it appears to be having issues
with source units not being expanded with proper values resulting in
syntax errors.

This also removes parser.adb from prepare (via config.linux) since it
was clobbering the file.

Recently sparforte is embedding timestamps in the source (sigh).  This
reimplements their logic but accepting SOURCE_DATE_EPOCH if available.