Move launcher from heredoc to repository and include relevant warning message
The warning message doesn't do much good in build(), the package may not
be installed on the system that built it, and it's likely to get lost.
Moving it to the launcher (where the same test was being run anyway)
and outputting it on STDERR makes the whole thing a bit less of a monkey
patch. Also this makes reviewing changes to the AUR package a lot
easier, changes to the launcher can be easily reviewed in git.