Switch back to use gmmlib source due to upstream reverting changes
Ustream was previously requiring gmmlib source (and providing
gmmlib in package) in order to build, then it switched to require
gmmlib installed on system and then it switched back to require
gmmlib sources due to regressions.

In the future, upstream intends to switch to require gmmlib
installed on system as the final solution, so be prepared for
another change(s) in this approach some upstream commits ahead.

That's the cost that you pay for using a development (-git)
package ;)

References
----------
https://github.com/intel/media-driver/issues/299#issuecomment-418156001