Initial commit:
It took me a day and a half get this into AUR, including packaging and validating
depends that have depends upon depends, then filing bug reports upstream,
and I even had to fork this because upstream no longer works. But this works, mostly.

As of now this is marching off my fork that contains bug fixes, and work arounds