propagate upstream update
The changes in this version deal with problems stemming from
modifications done to the Ubuntu version of gnome-shell. These are
either introduced by patches only included in the Ubuntu variant or
incompatibilities to other extensions Ubuntu includes by default. As
some Arch Users may want to use the Ubuntu version and compatibility
with other extensions is improved, the changes are worth shipping out to
them, too.

Version 59 was again withdrawn from review by the author, jumping
directly to 60.