Update conflicts (add old package names)
gmmlib-git was renamed to intel-gmmlib-git to follow the
name that is taken in the official [community] repository.

The change in this commit will properly remove gmmlib/gmmlib-git
from system if the user is installing the new package when having
the old one installed (to clarify, when upgrading from the old to
the new package name).