Match changes of the corresponding repository package
intel-media-driver and intel-gmmlib (formerly named gmmlib)
packages are now on the [community] official repository.
This commit will match the changes that were introduced
in the corresponding repository packages, as follows:

- pkgdesc: includes igpu target specification.
- depends: update gmmlib-git package name to reflect the new one
           (intel-gmmlib-git).
- Do not install the profile.d file. It means that the package do
  not to offer anymore a way to select the iHD driver. User needs
  to read the Arch Wiki and do their homework to select the iHD
  driver (Wiki page: currently at 'Hardware video acceleration').

Note:
I have still kept the 'YYYY.x.x' pkgver format instead of adopting
the new 'YY.x.x' one that is being currently used by upstream (and
was adopted by the repository package). Upstream have made a mess
with the versioning schema/method, altering it at each new major
release. So I will be still waiting for '19.x.x' to see if they
will keep it or change it again. Maybe I don't even change it at
all in the future (this will not interfere with the corresponding
repository package anyway). This also removes the need of adding
an epoch number at this moment.

Other changes in this commit:
- pkgver(): cosmetic changes and better handling of year/major.