Update to version 2.0.1
The build scripts were split out from the main repository into the
wp-cli-bundle repository, which pulls in the main repository and all
command repos as dependencies.

The bash completion file still lives in the main repo, and we still
pull that down as a separate source file.

We're using the composer script to run phpunit now, and it appears
to work with PHP 7.2 just fine, so the logic around that has been
removed.