Move Picard version check from $conflicts to $depends
$conflicts is for packages that have conflicting files, and Picard
hasn't had files in the same location as the plugins for a long time
now. Instead, let's make sure that the user is on a recent version of
Picard by checking the version in $depends.

(Even though we're on Arch, so only ever the latest stable version is
supported, and 2.0.x has been in Arch for a while now, so we shouldn't
even need to check this…)