Adds a pacman package definition for growpart (part of cloud-utils).
This package has a minimal dependency footprint as it includes growpart
only (rather than the entire cloud-utils suite).  It also includes a
wrapper script 'growpartfs' to grow both a partition and the file system
in it, and a systemd service that can be used to automatically grow a
file system on boot.