Remove incorrect hostid munging.
This is guaranteed to ensure the hostid gets quickly set to a single,
statically allotted hostid, then immediately reset back to the default
as defined by glibc's gethostid() call (which itself is statically based
on your IP address, usually 127.0.0.1).

If your system is one of the many systems on which the hostid is
007f0101, this won't fix that, and packaging the file means that anyone
who previously had this set by the administrator, will get file
conflicts.