upgpkg: zfs-dkms 0.7.12-1
upstream release

Move versioned spl and utils dependencies to package() in order to allow
the package to build with mismatched dependencies. In order to install
the resulting packages, the dependent packages will need to be installed
in the same transaction, but without the build-time dependency, it is no
longer necessary to uninstall the package set for as a prerequisite to
building the updated packages.