explain patching of older versions
Waf is also intended to be compiled into a binary for project-local use.
A change in Python breaks these binaries when they are created from a
waf version before 2.0.7. Thus, an explanation on how to recompile and
patch these enables maintainers unable to upgrade to the latest waf
right away to still build their projects.