1.4.4-2: fix emoji replacement, now using redesigned Noto Color Emoji
In 1.4.3, Telegram heavily refactored how they handle the emoji
spritemaps. Now they only include one size and scale/cache it
internally, and the sprite maps is broken into five 32x16 sheets.

Because of this, I had to actually learn how to generate the sprite
sheets, which was an adventure that's documented on
https://github.com/aswild/telegram-emoji-gen.

The latest commits in the noto emoji repository are the redesigned style
for Unicode 11, and there were a lot of SVGs missing in the older
commits.  Presumably there's some magic way to extract them from
a compiled font, but that's not something I have the willpower to dig
into right now.  The flags are also ugly and not stylized, but oh well.