Temporarily disable libsmbclient due to upstream bug
ffmpeg fails to compile with the newly released smbclient 4.9[1].
Currently, there is a pull request[2] at upstream samba that
aparently fixes this, but it still was not accepted at the time
of writing. It is being discussed on the samba mailing list[3].
Disabling libsmbclient until there is a consistent upstream
solution.

References
----------
[1] https://bugs.gentoo.org/666548
[2] https://github.com/samba-team/samba/pull/212/
[3] https://lists.samba.org/archive/samba-technical/2018-October/130668.html