Updated to latest upstream version.
Ok, TI updated their release to have the latest headers and linker scripts as a
separate file. It looks like it was an oversight on their part. We now have
their latest update on the PKGBUILD at least until they once again update their
stuff and tell nobody in the process.