Rework permissions and ownership for movim
This change is likely to break current setups.

The default user is now called "movim", created with the sysuser
configuration file, to avoid sharing ownership with other applications
(that would use http).

This will require a change in your php-fpm or apache setup. You should
change it to use "movim".

```
user = movim
group = http
```

Also, symlinks for the "cache" and "users" folders that were previously
created with the install script, (and thus not owned by the package),
are now created directly in the package.

You will be told by pacman that these files exist and the package can't
be installed. Please remove them before installing the new package.

Specifically, the two following symlinks:
/usr/share/webapps/movim/cache and
/usr/share/webapps/movim/users