Bump package release version to 0.12.1.2-2
Moves libpod-networking into main libpod package. libpod-networking
was an optional dependency when petris first created the libpod AUR
package, but it was eventually changed to be a hard dependency. Now
there is no point having it as a separate "package" and we can just
move it into the main libpod package.