Set API_ID and API_HASH
Telegram requires to set API_ID and API_HASH but do not provide
a way to receive a pair.

The official API_ID seems to be 2040 while the API_HASH is
"b18441a1ff607e10a989891a5462e627".

Use the defaults for now but might at some point use the official ones
as noted above.

[1]: https://github.com/telegramdesktop/tdesktop/commit/65b2db216033aa08b7bc846df27843e566f08981
[2]: https://github.com/telegramdesktop/tdesktop/issues/4717