upgpkg: rocketchat-desktop-2.14.6-2
The package now uses the system electron2. Therefore it is now a
dependency of this package. Don't worry about the additional size of the
new dependency, because this package lost most than 120M of size itself!

Spell checking now uses the system hunspell dictionaries instead of using the
ones provided by upstream. This is a breaking change and you need to
install the desired hunspell-* packages to re-enable spell checking.
Again, this is a zero sum game, the size gained by installing system
hunspell dicts is lost through removing the package dicts.