Disable plugins support
Plugins don't work.  One issue was the missing "lib/ghc-8.6.3/settings",
caused by the use of a throwaway GHC installation.  This could be
mitigated by switching to a permanent GHC installation (e.g. by creating
a separate AUR package for GHC pinned to 8.6).

However, plugins in Gitit are still not usable because they cannot be
compiled without libraries.  Bundling libraries with this package isn't
a good idea because

 1. it can only be achieved by turning this into a binary package due to
    limitations of Haskell's build tools, and
 2. even if problem 1 is fixed, the plugins are limited to just the
    bundled libraries.

Remark on problem 1: It is necessary have to bundle the gitit libraries
and all its dependencies.  makepkg requires packages to be built in a
"staged" directory ($pkgdir) that isn't the final destination of these
files.  Since cabal sandboxes are not relocatable ( see for example
https://github.com/haskell/cabal/issues/4097 ), nor does cabal support
building in staged sandboxes, the only solution would be to package
gitit as a binary tarball, which is undesirable.

As a result, plugin support will be disabled entirely, which shaves off
like 60MiB of space.  Users who wish to use gitit with plugins should
consider building and running gitit with cabal + ghc-pristine.  For
example:

    sudo pacman -S --needed cabal-install ghc-static git mailcap zlib
    curl -L https://aur.archlinux.org/cgit/aur.git/snapshot/ghc-pristine.tar.gz | tar -xz
    (cd ghc-pristine && makepkg -i)

    git clone https://github.com/jgm/gitit
    cd gitit
    cabal v2-update
    cabal v2-configure --with-compiler=/usr/share/ghc-pristine/bin/ghc
    cabal v2-run -- gitit -f example.conf