katriawm: Adjust build process
SolarAquarion made a note about the missing man pages.
Looking in the generated package the issue was more grave. The whole PREFIX thing while invoking make didn't work apparently, as the stuff was installed to /usr/local/
Switched to a patch file which adjusts the config.mk which gets sourced by the Makefile.