bump to 1.5.8-1
It's been fun, but this is probably the last version of
telegram-desktop-systemqt-notoemoji.

As of version 1.5.8, Telegram Desktop supports changing the emoji set,
basically rendering this custom package irrelevant.

In 1.5.8 you can choose between Mac (default), Android, Twemoji, and
EmojiOne. And unlike my janky scripts, the flags in Android emoji
actually look correct in the official version.

As much as I enjoyed maintaining this package and playing around with
its builds, there's no longer any benefit here compared to the
telegram-desktop package in [community].

Signed-off-by: Allen Wild <allenwild93@gmail.com>