Tweak source URL
The semantics of the "local" scheme are not well defined. However, it's
wise to use the scheme with a forward-looking attitude, so that it will
remain compatible with whatever specification might be put in place in
the future.

The best analog to "local" is "file," for which all of the following are
valid and equivalent:

    file:///absolute/path
    file://localhost/absolute/path
    file:/absolute-path

The important bit here is that "//" precedes the authority (hostname).
Do the same thing for this package.